+++ src/components/layout/Header.tsx
@@ -48,11 +48,23 @@ export default function Header() {
 
 	const {
 		search: { searchActive, results },
+		searchDispatch,
 	} = useContext(SearchContext);
 
 	const { loggedInId } = useViewer();
 
 	const [isLargerThanMd] = useMediaQuery('(min-width: 48rem)');
+
+	const handleDrawerOpen = () => {
+		drawerOnOpen();
+
+		searchDispatch({
+			type: 'OPEN_SEARCH',
+			payload: {
+				searchDrawerClose: drawerOnClose,
+			},
+		});
+	};
 
 	const handleLogout = () => {
 		logoutMutation().then(() => {
@@ -63,7 +75,7 @@ export default function Header() {
 	const SearchButton = () => (
 		<Button
 			ref={drawerButtonRef}
-			onClick={drawerOnOpen}
+			onClick={handleDrawerOpen}
 			leftIcon={isLargerThanMd ? <FiSearch /> : undefined}
 			aria-label='Search for candidates'
 			colorScheme='gray'
@@ -222,24 +234,9 @@ export default function Header() {
 									</LightMode>
 								</>
 							) : (
-								<>
-									<Button
-										as={RouterLink}
-										to='/register'
-										borderRadius={{ base: 'full', md: 'lg' }}
-										bgColor='brand.orange'
-										_hover={{
-											bgColor: 'orange.300',
-										}}
-										color='text.dark'
-										size='lg'
-									>
-										Join RISE
-									</Button>
-									<Link as={RouterLink} to='https://risetheatre.org' my={0} isExternal>
-										<Image src={circleLogo} alt='RISE icon' loading='eager' h='100px' py={4} />
-									</Link>
-								</>
+								<Link as={RouterLink} to='https://risetheatre.org' my={0} isExternal>
+									<Image src={circleLogo} alt='RISE icon' loading='eager' h='100px' py={4} />
+								</Link>
 							)}
 						</Stack>
 					</Container>