+++ src/routes/Help.tsx
@@ -1,13 +1,10 @@
 import Page from '../components/Page';
-import { usePostContent } from '../hooks/queries/usePostContent';
-import { Container, Spinner } from '@chakra-ui/react';
+import ContentView from '../views/ContentView';
 
 export default function Help() {
-	const [content, { contentLoading, contentError }] = usePostContent('959');
-
 	return (
 		<Page title='Help'>
-			<Container maxW='4xl'>{contentLoading && !contentError ? <Spinner /> : content}</Container>
+			<ContentView postId='959' />
 		</Page>
 	);
 }