+++ src/views/SearchWizardView.tsx
@@ -1,63 +1,44 @@
 import React, { useContext } from 'react';
-import { Flex, Button, Stack } from '@chakra-ui/react';
+import { Box, Fade, Spacer } from '@chakra-ui/react';
 
 import SearchFilterDepartment from '../components/SearchFilterDepartment';
 import SearchFilterJobs from '../components/SearchFilterJobs';
 import SearchFilterSkills from '../components/SearchFilterSkills';
 
 import { SearchContext } from '../context/SearchContext';
+import SearchFilterName from '../components/SearchFilterName';
+import AdditionalSearchFilters from '../components/AdditionalSearchFilters';
 
 interface Props {
 	showButtons?: boolean;
 	onSubmit: (e: React.FormEvent<HTMLFormElement>) => void;
 }
 
-export default function SearchWizardView({ showButtons, onSubmit }: Props) {
+export default function SearchWizardView({ onSubmit }: Props) {
 	const {
 		search: {
 			filters: {
+				name,
 				positions: { department, jobs },
 			},
-			searchActive,
 		},
-		searchDispatch,
 	} = useContext(SearchContext);
 
-	const handleReset = () => {
-		searchDispatch({
-			type: 'RESET_SEARCH_FILTERS',
-			payload: {},
-		});
-	};
-
 	return (
-		<form id='search-candidates' onSubmit={onSubmit}>
-			<Stack direction='column' gap={8}>
-				{/* TODO convert Department, Jobs, and Skills to unified <CheckboxButtonFilterGroup> interface */}
-				{/* Step 1 */}
-				<SearchFilterDepartment heading='Which department are you hiring for?' />
+		<>
+			<SearchFilterName />
 
-				{/* Step 2 */}
-				{department ? <SearchFilterJobs heading='What job(s) are you looking to fill?' /> : null}
-
-				{/* Step 3 */}
-				{department && jobs && jobs.length > 0 ? (
-					<SearchFilterSkills heading='What skills are you looking for?' />
-				) : null}
-
-				{showButtons ? (
-					<Flex gap={2}>
-						{searchActive ? (
-							<Button type='submit' size='lg'>
-								Search
-							</Button>
-						) : null}
-						<Button type='reset' size='lg' onClick={handleReset}>
-							Reset
-						</Button>
-					</Flex>
-				) : null}
-			</Stack>
-		</form>
+			<Fade in={!name}>
+				<form id='search-candidates' onSubmit={onSubmit}>
+					<Box my={4} height={name ? 0 : 'auto'}>
+						<SearchFilterDepartment />
+						{department ? <SearchFilterJobs /> : null}
+						{department && jobs && jobs.length > 0 ? <SearchFilterSkills /> : null}
+						<Spacer h={8} />
+						<AdditionalSearchFilters />
+					</Box>
+				</form>
+			</Fade>
+		</>
 	);
 }