+++ changelog.md
@@ -2,6 +2,10 @@
 
 All notable changes to this project will be documented in this file, starting with 1.0.0beta.
 
+## [unreleased]
+
+- Added: ContentView component for displaying WP content
+
 ## [1.0.1beta] - 2023-05-16
 
 - Added: Google reCAPTCHA v3 integration